obd-logger/requirements.txt
python-can
fastapi
uvicorn[standard]
lttbc
numpy
